feat: add (explicit) zstd compression extensions
After discussion we've decided that we want to have a compression
extension that is really easy to get up and running with, and
doesn't require special packaging considerations. The current
implementation, based on `zstd`, requires a binary dependency
and special packaging considerations because `zstd` is not
supported in the node.js stdlib. Therefore, we are going to
migrate the "default" compression extensions to use gzip, which
is supported in the stdlib and thus won't require any special
considerations for packaging.

This commit is the first step toward that; in this commit we
simply make a copy of the existing compression extensions and
change the package name to include a `-zstd` suffix. Once this
package is deployed to npmjs, we can migrate existing users
to it and then re-purpose the original one.