feat: Switch default compression to gzip (#1267)
* feat: Switch default compression to gzip

Now that we have the @gomomento/sdk-nodejs-compression-zstd package,
this commit updates the default compression extensions package
to use gzip as the compression algorithm rather than zstd. This
allows us to get rid of the binary dependency that makes packaging
complicated. We made the decision to do this after benchmarking
showed very little difference in compression ratio and only a little
difference in performance between the two compression libraries;
therefore, shedding the packaging complexity for a simpler user
experience seemed like the best UX choice.