fix: remove `TEST_CACHE_NAME` env var reference in tests
Previoulsy we read the test cache name from the environment in CI. We
no longer do this. However a developer may have `TEST_CACHE_NAME`
set.

In that case the integration tests will always use that for the cache
name. Because of this, control tests fail since some invoke
`testCacheName()` expecting a unique cache name as opposed to a
consistent one.

Speaking with other developers, they have been hit by this gotcha. One
solution would be to change tests to specify if they expect a unique
cache name vs some cache name. That will make the tests more difficult
to maintain, so we will remove the env var instead.