chore: port initial batch of example code to use `switch` style (#1342)
* chore: port initial batch of example code to use `switch` style

This commit ports the first handful of API examples over to use
the new switch-based pattern matching style. It also updates
the API reference docs for `get` and `getBatch` to show both
the "simplified" and "pattern matching" styles of accessing
the retrieved data.

This is intended to serve as a starting point for discussion
about whether we like this approach for the docs. I built the docs
locally and looked at these examples, and I like them. If other
folks on the team are on board with this approach then we can
move forward quickly on porting the rest of the docs/examples
over.

---------

Co-authored-by: Matt Straathof <matt.straat@gmail.com>