fix: broken auth tests
This commit fixes the auth tests; we no longer use the auth0 token
for instantiating any clients, we only use the MGA account session
token.

This commit also changes the run scripts in the `package.json` such
that the generic `integration-test` target runs *all* integration
tests; previously it was only running cache/control/topics tests,
which felt counter-intuitive to me. I added a new target that runs
only the cache/control/topics tests but it seemed to me like the
default behavior should be to run all of the tests, to avoid
surprising anyone and missing the execution of certain tests.

There may be changes necessary in the canaries code to account for this.