refactor: organize integration tests by service (#1410)
Organizes the integration tests by service, and importantly by directory. By organizing by directory, we have an unambiguous regex to use for jest to match integration tests by service, eg jest cache/ to run all cache service tests. Integration test authoring stays maintainable because developers just need to place their integration test file under the appropriate service directory.

For clarity we have refactored common-integration-tests, the integration tests under node-js, and those under web to have the same directory structure.

We have updated the scripts in the package.json files to match on the service (as a directory) as opposed to individual test files.

A note on the organization:

We moved the webhooks test under the topics service test directory.
We kept the client constructor tests in the top-level integration test directory.
In a following PR we will introduce makefiles that will invoke these scripts and update GitHub Actions. That is, the makefiles will run npm run integration-test-<>. That way the source of truth for the tests will be the scripts in package.json, since various shell scripts still reference these.