chore: simplify header/retry interceptor construction, collapse timeout interceptor
This commit does the following:

* Simplify the logic for constructing header/retry interceptors; previously it involved
  instantiating a class that didn't really need to exist. These are now simpler factory
  functions with less state involved.
* Collapse the timeout interceptor into the retry interceptor. Because the retry interceptor
  now needs to support overwriting the deadline for a given request in order to implement
  the more advanced type of retries that we want to support, it was starting to have overlaps/
  conflicts with the ClientTimeoutInterceptor. In this commit we simply refactor the basic
  overall-request-timeout logic into the RetryInterceptor and update all of the call sites to
  use RetryInterceptor instead of the old ClientTimeoutInterceptor.

There will still be another commit necessary in order to implement the intermediate timeouts that
were implemented by `RetryUntilTimeoutInterceptor`, but as of this commit we should have parity
with the previous support for retries and so integration tests should pass.

client-timeout-interceptor.ts is left in place, commented out for now. Will be removed in a
subsequent commit once the logic is ported into the retry interceptor.