chore: adding examples that are currently hard-coded in dev docs
This commit copies over a lot of example JS code that is currently
hard-coded in the dev docs, into example files here that we can
validate in CI. Also hopefully makes it easier to make sweeping
changes in the future (such as our instanceof migration).

A subsequent dev docs PR will consume these and get rid of the
hard-coded examples.