fix: copy constructors should use spread operator (#1494)
In order to prevent bugs in our config copy constructors, we should
use the spread operator over `this`.

Previously in the copy constuctors we built a copy of the
corresponding properties by hand. Because some properties are
optional, and because we add new properties over time, we introduced
drift between the new properties and various copy constructors.

That is, when we added new properties, some copy constructors were out
of date and hence falling back to defaults instead of copying those
new properties over.

We can fix this across the board by:
1. aligning the field names in each config class with the corresponding
property name, and
2. building the props copy using a spread over `this` followed by the
new value.