chore: bump zstd version (#1504)
Bumps the zstd version in the zstd compression library to 2.0.0. Though this version maintains API compatibility with version 1.0, the binary may be slightly different. Hence we update the tests. Additionally, version 2.0 does not throw an error when data is compressed but the frame header is invalid. Instead it returns a 0-length byte array. We remove that test.