fix: aws lambda zstd example needs postinstall step
The current version of `@mongodb-js/zstd` has a deficiency loading the
native node module as compared to the previous major version. The
loader needs fallbacks to
`@mongodb/zstd/build/{Release,Debug}/zstd.node` to fully work when
bundling for lambda. We will open an issue with mongo about this.

In the interim we add a postinstall step to copy the file in the
`Debug` directory so the build will pass.