fix: aws lambda zstd example needs postinstall step (#1529)
The current version of @mongodb-js/zstd has a deficiency loading the
native node module as compared to the previous major version. The
loader needs fallbacks to
@mongodb/zstd/build/{Release,Debug}/zstd.node to fully work when
bundling for lambda. We will open an issue with mongo about this.

In the interim we add a postinstall step to copy the file in the
Debug directory so the build will pass.

We also provide scripts to build and package for aws npm run rebuild:aws and dev npm run rebuild:dev.