fix: add bookkeeping for number of active subscriptions (#1564)
Found that the nodejs sdk does not keep track of the number of concurrent subscriptions while auditing the implementations of separate grpc channels for publish and subscribe.

This PR adds the necessary bookkeeping so that the TopicClient can reject new subscribe requests if all grpc channels are currently occupied with active streams.

Also added a new test file that can be run locally to verify that the TopicClient does not allow >100 streams on a single grpc channel and that it is capable of handling bursts of subscribe and unsubscribe requests.